Initial setup for Bubble Tea based app
A main.go file to act as root entrypoint of application. A cmd package
is set up with Cobra library to handle the sub commands, arguments, and
flags the user might need to provide when starting the program.

The ui package sets up the Bubble Tea program that is responsible for
rendering the application, handling state, etc. The package is split
into separate files to handle different aspects of the UI like the
organizations and projects lists. Later ui concerns like
rendering cloud run services, proxy servers and their logs will also be
handled in this package.

Signed-off-by: Snorre Magnus Davøen <post@snorre.io>